Move app command files into commands
