+++ commands/hub/_invite.js
