+++ commands/maps/_arenas.js
