+++ commands/maps/_battle-royale.js
